Adds some javadocs, moves some things around.
