+++ src/main/java/net/fabricmc/loader/api/config/serialization/AbstractTreeSerializer.java
@@ -18,7 +18,6 @@ package net.fabricmc.loader.api.config.serialization;
 
 import net.fabricmc.loader.api.config.ConfigDefinition;
 import net.fabricmc.loader.api.config.ConfigManager;
-import net.fabricmc.loader.api.config.ConfigSerializer;
 import net.fabricmc.loader.api.config.data.DataType;
 import net.fabricmc.loader.api.config.value.ValueContainer;
 import net.fabricmc.loader.api.config.value.ValueKey;
