+++ src/main/java/net/fabricmc/loader/api/config/serialization/TOMLSerializer.java
@@ -4,7 +4,6 @@ import net.fabricmc.loader.api.SemanticVersion;
 import net.fabricmc.loader.api.VersionParsingException;
 import net.fabricmc.loader.api.config.ConfigDefinition;
 import net.fabricmc.loader.api.config.ConfigManager;
-import net.fabricmc.loader.api.config.ConfigSerializer;
 import net.fabricmc.loader.api.config.data.DataType;
 import net.fabricmc.loader.api.config.exceptions.ConfigSerializationException;
 import net.fabricmc.loader.api.config.serialization.toml.Toml;
