+++ src/main/java/net/fabricmc/loader/config/ValueContainerProviders.java
@@ -16,7 +16,7 @@
 
 package net.fabricmc.loader.config;
 
-import net.fabricmc.loader.api.config.SaveType;
+import net.fabricmc.loader.api.config.data.SaveType;
 import net.fabricmc.loader.api.config.value.ValueContainer;
 import net.fabricmc.loader.api.config.value.ValueContainerProvider;
 import org.jetbrains.annotations.NotNull;
